Revert "[TMEM] Remove Unneeded Stores (#6892)"
This reverts commit 8297f8ccdb2d6a7fe314408d8299f1e49244cb3e.
